% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.Overlap.R
\name{Correlation.Overlap}
\alias{Correlation.Overlap}
\title{Correlation.Overlap}
\usage{
Correlation.Overlap(Data, Overlap, method)
}
\arguments{
\item{Data}{data.frame containing the expression data. Nodes on the Rows, Individuals on the Columns. Don't forget to give the names to the Nodes and to the Individuals. Nodes must have the row.names() with the Node Name.}

\item{Overlap}{A vector containg the names of the Nodes of interest.}

\item{method}{Spearman ("s", "spearman") or Pearson ("p", "pearson") correlation}
}
\description{
This function computes the correlation between Nodes and the Overlapping Nodes of interest.
}
