% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampledfExpression.R
\docType{data}
\name{ExampledfExpression}
\alias{ExampledfExpression}
\title{ExampledfExpression}
\format{data.frame 268 lines, 18 columns.}
\usage{
ExampledfExpression
}
\description{
ExampledfExpression data.frame containing expression data for 1000 genes and 18 individuals.
}
\keyword{datasets}
