% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleGRF.R
\docType{data}
\name{ExampleGRF}
\alias{ExampleGRF}
\title{ExampleGRF}
\format{
data.frame 184 lines, 1 column.
}
\usage{
data(ExampleGRF)
}
\description{
ExampleGRF data.frame containing data.frame containing names of GRFs.
}
\keyword{datasets}
