% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomics.R
\docType{data}
\name{metagenomics_abundance}
\alias{metagenomics_abundance}
\title{metagenomics_abundance}
\format{
data.frame from The USC Microbial Observatory. The data is public available at <https://www.ebi.ac.uk/metagenomics/projects/ERP013549>
}
\usage{
data('metagenomics_abundance')
}
\description{
metagenomics_abundance
}
\keyword{datasets}
