% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampledfExpression.R
\docType{data}
\name{Microarray_Expression1}
\alias{Microarray_Expression1}
\title{Microarray_Expression1}
\format{
data.frame 1000 lines, 18 columns.
}
\usage{
Microarray_Expression1
}
\description{
Microarray_Expression1 data.frame containing expression data for 1000 genes and 18 individuals.
}
\keyword{datasets}
