% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampledfExpression2.R
\docType{data}
\name{Microarray_Expression2}
\alias{Microarray_Expression2}
\title{Microarray_Expression2}
\format{
data.frame 1000 lines, 18 columns.
}
\usage{
Microarray_Expression2
}
\description{
Microarray_Expression2 data.frame containing expression data for 1000 genes and 18 individuals.
}
\keyword{datasets}
