% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmcor.R
\name{rmcor}
\alias{rmcor}
\title{wTO.aux.each}
\usage{
rmcor(ID, Measure1, Measure2)
}
\description{
Computes the repeated measures correlation. based on Blendman Altman (1995). Implemented by <https://doi.org/10.3389/fpsyg.2017.00456>
}
\keyword{internal}
