% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ind.R
\name{sample_ind}
\alias{sample_ind}
\title{sample_ind}
\usage{
sample_ind(x, dfExpression)
}
\arguments{
\item{x}{Column to be resampled}

\item{dfExpression}{data.frame object containing the genes expression on the rows and the individuals (Individuals in the Columns)}
}
\description{
Computes the resufling of the expression values for the IDs.
}
\author{
Deisy Morselli Gysi <deisy at bioinf.uni-leipzig.de>
}
\keyword{internal}
