% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wTO.aux.each.R
\name{wTO.aux.each}
\alias{wTO.aux.each}
\title{wTO.aux.each}
\usage{
wTO.aux.each(n, Data, Overlap, method, method_resampling, lag, ID)
}
\arguments{
\item{n}{Number of bootstraps / reshuffles to be run for the estimates of the "Threshold" or "pval".}

\item{Data}{data.frame containing the count / expression data for the correlation.}

\item{Overlap}{Nodes of interested, where the Overlapping weights will be computed.}

\item{method}{Type of the correlation that should be used. "s" / "spearman" will compute the rank spearman correlation, "p" / "pearson" will compute the linear correlation. If no value is given, the default is to use "s".}

\item{method_resampling}{method of the resampling. Bootstrap or Reshuffle. Bootstrap null hypothesis is that the wTO is random, and Reshuffle tests if the wTO is equal to zero.}
}
\description{
wTO.aux.each calculate the wTO for each one of the resamplings.
}
\keyword{internal}
