% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wTO.in.line.R
\name{wTO.in.line}
\alias{wTO.in.line}
\title{wTO.in.line}
\usage{
wTO.in.line(d)
}
\arguments{
\item{d}{correlation matrix to be converted into the line format.}
}
\value{
the wTO matrix into a data.frame: Node1, Node2 and wTO.
}
\description{
Transforms a correlation matrix into the line format.
}
\author{
Deisy Morselli Gysi <deisy at bioinf.uni-leipzig.de>
}
