% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wTO.R
\name{wTO}
\alias{wTO}
\title{wTO}
\usage{
wTO(A, sign = c("abs", "sign"))
}
\arguments{
\item{A}{Is the weighted adjency matrix (correlation matrix).}

\item{sign}{("abs", "sign") if the user wants to use the absolute correlation or the signed correlation.}
}
\value{
A matrix containing the wTO values.
}
\description{
Calculates the weighted topologycal overlap (wTO)
between a set of Nodes and the Overlapping nodes. This function implements the method from \cite{nowick2009differences}.
}
\references{
\url{http://www.pnas.org/content/106/52/22358.full.pdf}
}
