% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wTO.rep_measure.R
\name{wTO.rep_measure}
\alias{wTO.rep_measure}
\title{wTO.rep_measure}
\usage{
wTO.rep_measure(
  Data,
  Overlap = row.names(Data),
  ID,
  sign = "sign",
  delta = 0.2,
  n = 10
)
}
\arguments{
\item{Data}{data.frame containing the count / expression data for the correlation.}

\item{Overlap}{Set of nodes of interest, where the Overlapping weights will be computed.}

\item{ID}{a vector with the individuals identification}

\item{sign}{Should the wTO be signed?}

\item{delta}{expected difference between the real wTO and the bootstraped.}

\item{n}{Number of resamplings, used to compute the empirical distribuitions of the links. Default is set to 100.}
}
\description{
Compute the wTO for a repeated measures expermiment and also the bootstraps. Proposed at arXiv:1711.04702. This is a quicker version of the wTO.Complete. It doesn'T contain diagnose plots nor a parallel version.
}
\examples{

#wTO.rep_measure(Data = Microarray_Expression1, ID = rep(c(1:9),2), 
#Overlap = ExampleGRF$x)
}
\author{
Deisy Morselli Gysi <deisy at bioinf.uni-leipzig.de>
}
