% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.Overlap.R
\name{Correlation.Overlap}
\alias{Correlation.Overlap}
\title{Correlation.Overlap}
\usage{
Correlation.Overlap(Data, Overlap, method)
}
\arguments{
\item{Data}{data.frame containing the expression data. Genes on the Rows, Individuals on the Columns. Don't forget to give the names to the Genes and to the Individuals. Genes must have the row.names() with the Gene Name.}

\item{Overlap}{A vector containg the names of the Gene Regulatory Factors (GRFs).}

\item{method}{Spearman ("s", "spearman") or Pearson ("p", "pearson") correlation}
}
\value{
List containing
\item{Descriptive}{the quantiles of the correlation and mean.}
\item{Final_Correlation}{the correlation matrix where the diagonal is 0.}
}
\description{
This function computes the correlation between Nodes and the Overlapping Nodes of interest.
}
