% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cut.off2.R
\name{Cut.off}
\alias{Cut.off}
\title{Cut.off}
\usage{
Cut.off(wTO_value, type, plot)
}
\arguments{
\item{wTO_value}{is the table returned by the wTO with all the simulated values.}

\item{main}{title of the graph.}
}
\value{
plots in a new device the cutoff value, and the amount of links in each one of the bands.
}
\description{
Computes the cutoff threshold using the descriptive values generated by the bootsrap, also plots the empirical distribution and the sample wTO.
}
\author{
Deisy Morselli Gysi <deisy at bioinf.uni-leipzig.de>
}
\keyword{internal}
