% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wTO.R
\name{wTO}
\alias{wTO}
\title{wTO}
\usage{
wTO(A, sign = c("abs", "sign"))
}
\arguments{
\item{A}{Is the weighted adjency matrix (correlation matrix).}

\item{sign}{("abs", "sign") if the user wants to use the absolute correlation or the signed correlation.}
}
\value{
A matrix containing the wTO values.
}
\description{
Calculates the weighted topologycal overlap (wTO)
between a set of Nodes and the Overlapping nodes. This function implements the method from Nowick (2009).
}
\references{
Katja Nowick, Tim Gernat, Eivind Almaas and Lisa Stubbs (2009) <doi:10.1073/pnas.0911376106>
}
\author{
Deisy Morselli Gysi <deisy at bioinf.uni-leipzig.de>
}
